Organize admin tests and fix authed_only to return 403 on unauthed
