+++ CTFd/utils/decorators/__init__.py
@@ -74,7 +74,10 @@ def authed_only(f):
         if authed():
             return f(*args, **kwargs)
         else:
-            return redirect(url_for('auth.login', next=request.path))
+            if request.content_type == 'application/json':
+                abort(403)
+            else:
+                return redirect(url_for('auth.login', next=request.path))
 
     return authed_only_wrapper
