+++ tests/admin/test_views.py
@@ -68,151 +68,6 @@ def test_get_admin_as_user():
     destroy_ctfd(app)
 
 
-def test_get_admin_statistics():
-    app = create_ctfd()
-    with app.app_context():
-        register_user(app)
-        client = login_as_user(app, name="admin", password="password")
-        r = client.get('/admin/statistics')
-        assert r.status_code == 200
-    destroy_ctfd(app)
 
 
-def test_get_admin_notifications():
-    app = create_ctfd()
-    with app.app_context():
-        register_user(app)
-        client = login_as_user(app, name="admin", password="password")
-        r = client.get('/admin/notifications')
-        assert r.status_code == 200
-    destroy_ctfd(app)
 
-
-def test_get_admin_pages():
-    app = create_ctfd()
-    with app.app_context():
-        register_user(app)
-        client = login_as_user(app, name="admin", password="password")
-        r = client.get('/admin/pages')
-        assert r.status_code == 200
-    destroy_ctfd(app)
-
-
-def test_get_admin_pages_new():
-    app = create_ctfd()
-    with app.app_context():
-        register_user(app)
-        client = login_as_user(app, name="admin", password="password")
-        r = client.get('/admin/pages/new')
-        assert r.status_code == 200
-    destroy_ctfd(app)
-
-
-def test_get_admin_users():
-    app = create_ctfd()
-    with app.app_context():
-        register_user(app)
-        client = login_as_user(app, name="admin", password="password")
-        r = client.get('/admin/users')
-        assert r.status_code == 200
-    destroy_ctfd(app)
-
-
-def test_get_admin_users_new():
-    app = create_ctfd()
-    with app.app_context():
-        register_user(app)
-        client = login_as_user(app, name="admin", password="password")
-        r = client.get('/admin/users/new')
-        assert r.status_code == 200
-    destroy_ctfd(app)
-
-
-def test_get_admin_teams():
-    app = create_ctfd()
-    with app.app_context():
-        register_user(app)
-        client = login_as_user(app, name="admin", password="password")
-        r = client.get('/admin/teams')
-        assert r.status_code == 200
-    destroy_ctfd(app)
-
-
-def test_get_admin_teams_new():
-    app = create_ctfd()
-    with app.app_context():
-        register_user(app)
-        client = login_as_user(app, name="admin", password="password")
-        r = client.get('/admin/teams/new')
-        assert r.status_code == 200
-    destroy_ctfd(app)
-
-
-def test_get_admin_scoreboard():
-    app = create_ctfd()
-    with app.app_context():
-        register_user(app)
-        client = login_as_user(app, name="admin", password="password")
-        r = client.get('/admin/scoreboard')
-        assert r.status_code == 200
-    destroy_ctfd(app)
-
-
-def test_get_admin_challenges():
-    app = create_ctfd()
-    with app.app_context():
-        register_user(app)
-        client = login_as_user(app, name="admin", password="password")
-        r = client.get('/admin/challenges')
-        assert r.status_code == 200
-    destroy_ctfd(app)
-
-
-def test_get_admin_challenges_new():
-    app = create_ctfd()
-    with app.app_context():
-        register_user(app)
-        client = login_as_user(app, name="admin", password="password")
-        r = client.get('/admin/challenges/new')
-        assert r.status_code == 200
-    destroy_ctfd(app)
-
-
-def test_get_admin_submissions():
-    app = create_ctfd()
-    with app.app_context():
-        register_user(app)
-        client = login_as_user(app, name="admin", password="password")
-        r = client.get('/admin/submissions')
-        assert r.status_code == 200
-    destroy_ctfd(app)
-
-
-def test_get_admin_submissions_correct():
-    app = create_ctfd()
-    with app.app_context():
-        register_user(app)
-        client = login_as_user(app, name="admin", password="password")
-        r = client.get('/admin/submissions/correct')
-        assert r.status_code == 200
-    destroy_ctfd(app)
-
-
-def test_get_admin_submissions_incorrect():
-    app = create_ctfd()
-    with app.app_context():
-        register_user(app)
-        client = login_as_user(app, name="admin", password="password")
-        r = client.get('/admin/submissions/incorrect')
-        assert r.status_code == 200
-    destroy_ctfd(app)
-
-
-def test_get_admin_config():
-    app = create_ctfd()
-    with app.app_context():
-        register_user(app)
-        client = login_as_user(app, name="admin", password="password")
-        r = client.get('/admin/config')
-        assert r.status_code == 200
-    destroy_ctfd(app)
